job details api call done
