+++ app/(public)/jobs/details/[id]/page.tsx
@@ -1,0 +1,7 @@
+"use client";
+import JobDetail from "@/components/job/view";
+import React from "react";
+
+export default function Page() {
+  return <JobDetail/>;
+}
