+++ package.json
@@ -14,6 +14,7 @@
     "axios": "^1.12.0",
     "cookies-next": "^4.3.0",
     "date-fns": "^4.1.0",
+    "dayjs": "^1.11.18",
     "lucide-react": "0.474.0",
     "next": "^14.2.32",
     "react": "18.2.0",
